// src/app/modules/B2BSongs/B2BSongs.interface.ts
import mongoose from 'mongoose';

import { IUser } from '../user/user.interface';

export type IB2BSongs = {
  image: string;
  video: string;
  videoId: string;
  sourceCompany: string;
  durationMs: number;
  ddexS3Folder?: string;
  ddexStatus?: string;
  version:
    | 'Lyrical Video'
    | 'Interview'
    | 'Alternative Version'
    | 'Official'
    | 'Live'
    | 'Behind The Scenes'
    | 'Teaser'
    | 'Original Content'
    | 'Pseudo Video';
  explicit: 'Yes' | 'No';
  title: string;
  primaryArtist: string[];
  // primaryArtist: [mongoose.Types.ObjectId | IPrimaryArtist];
  featuringArtists: string[];
  writer: string;
  composer: string;
  musicDirector: string;
  producer: string;
  editor: string;
  label: string;
  // label: mongoose.Types.ObjectId | ILabel;
  user: mongoose.Types.ObjectId | IUser;
  genre: string;
  subGenre: string;
  language: string;
  upc: string;
  isrc: string;
  audioIsrc: string;
  storeReleaseDate: string;
  releaseDate: string;
  alreadyHaveAnVevoChannel: 'Yes' | 'No';
  videoAlreadyExistOnYoutube: 'Yes' | 'No';
  vevoChannel: string;
  videoLink: string;
  youtubeLink: string;
  assetId: string;
  keywords: [];
  copyright: string;
  copyrightYear: string;
  isKids: 'Yes' | 'No';
  isExist: 'Yes' | 'No';
  territoryPolicy: string;
  youtubePremiere: 'Yes' | 'No';
  description: string;
  corrections: [];
  isCorrection: boolean;
  repertoireOwner: string;
  visibility: string;
  time: string;
  countdownTheme: string;
  countdownLength: string;
  isApproved: 'approved' | 'rejected' | 'pending';
  videoStatus: 'take-down' | 'distribute' | 'none';
};